Moved all classes to their proper place, created helper class
